+++ projectroles/tests/test_remote_projects_api.py
@@ -16,6 +16,7 @@ from projectroles.models import (
     RemoteProject,
     RemoteSite,
     SODAR_CONSTANTS,
+    AppSetting,
 )
 
 from projectroles.remote_projects import RemoteProjectAPI
@@ -26,6 +27,7 @@ from projectroles.tests.test_models import (
     RemoteSiteMixin,
     RemoteProjectMixin,
     SodarUserMixin,
+    AppSettingMixin,
 )
 
 User = auth.get_user_model()
@@ -88,6 +90,9 @@ PEER_SITE_USER_DISPLAY = True
 NEW_PEER_NAME = PEER_SITE_NAME + ' new'
 NEW_PEER_DESC = PEER_SITE_DESC + ' new'
 NEW_PEER_USER_DISPLAY = not PEER_SITE_USER_DISPLAY
+
+PROJECTROLES_APP_SETTING_IP_RESTRICT_UUID = str(uuid.uuid4())
+PROJECTROLES_APP_SETTING_IP_ALLOWLIST_UUID = str(uuid.uuid4())
 
 
 class TestGetTargetData(
@@ -184,6 +189,7 @@ class TestGetTargetData(
                 }
             },
             'peer_sites': {},
+            'app_settings': {},
         }
 
         self.assertEqual(sync_data, expected)
@@ -233,6 +239,7 @@ class TestGetTargetData(
                     'user_display': self.peer_site.user_display,
                 }
             },
+            'app_settings': {},
         }
 
         self.assertEqual(sync_data, expected)
@@ -285,6 +292,7 @@ class TestGetTargetData(
                 },
             },
             'peer_sites': {},
+            'app_settings': {},
         }
 
         self.assertEqual(sync_data, expected)
@@ -354,6 +362,7 @@ class TestGetTargetData(
                     'user_display': self.peer_site.user_display,
                 }
             },
+            'app_settings': {},
         }
 
         self.assertEqual(sync_data, expected)
@@ -412,6 +421,7 @@ class TestGetTargetData(
                 },
             },
             'peer_sites': {},
+            'app_settings': {},
         }
 
         self.assertEqual(sync_data, expected)
@@ -420,7 +430,12 @@ class TestGetTargetData(
         """Test get data with no project access set in the source site"""
         sync_data = self.remote_api.get_target_data(self.target_site)
 
-        expected = {'users': {}, 'projects': {}, 'peer_sites': {}}
+        expected = {
+            'users': {},
+            'projects': {},
+            'peer_sites': {},
+            'app_settings': {},
+        }
 
         self.assertEqual(sync_data, expected)
 
@@ -432,6 +447,7 @@ class TestSyncSourceData(
     RemoteSiteMixin,
     RemoteProjectMixin,
     SodarUserMixin,
+    AppSettingMixin,
     TestCase,
 ):
     """Tests for the sync_source_data() API function"""
@@ -441,6 +457,7 @@ class TestSyncSourceData(
         self.admin_user = self.make_user(settings.PROJECTROLES_DEFAULT_ADMIN)
         self.admin_user.is_staff = True
         self.admin_user.is_superuser = True
+        self.maxDiff = None
 
         # Init roles
         self.role_owner = Role.objects.get_or_create(name=PROJECT_ROLE_OWNER)[0]
@@ -515,6 +532,28 @@ class TestSyncSourceData(
                     'user_display': PEER_SITE_USER_DISPLAY,
                 }
             },
+            'app_settings': {
+                PROJECTROLES_APP_SETTING_IP_RESTRICT_UUID: {
+                    'name': 'ip_restrict',
+                    'type': 'BOOLEAN',
+                    'value': False,
+                    'value_json': {},
+                    'app_plugin_id': None,  # None is for 'projectroles' app
+                    'project_uuid': SOURCE_PROJECT_UUID,
+                    'user_id': None,
+                    'local': False,
+                },
+                PROJECTROLES_APP_SETTING_IP_ALLOWLIST_UUID: {
+                    'name': 'ip_allowlist',
+                    'type': 'STRING',
+                    'value': '',
+                    'value_json': {},
+                    'app_plugin_id': None,  # None is for 'projectroles' app
+                    'project_uuid': SOURCE_PROJECT_UUID,
+                    'user_id': None,
+                    'local': False,
+                },
+            },
         }
 
     def test_create(self):
@@ -539,6 +578,7 @@ class TestSyncSourceData(
         self.assertEqual(User.objects.all().count(), 2)
         self.assertEqual(RemoteProject.objects.all().count(), 3)
         self.assertEqual(RemoteSite.objects.all().count(), 2)
+        self.assertEqual(AppSetting.objects.count(), 2)
 
         new_user = User.objects.get(username=SOURCE_USER_USERNAME)
 
@@ -653,6 +693,75 @@ class TestSyncSourceData(
         peer_project_dict.pop('level')
         peer_project_dict.pop('date_access')
         self.assertEqual(peer_project_dict, expected)
+
+        app_setting_ip_restrict_obj = AppSetting.objects.get(
+            sodar_uuid=PROJECTROLES_APP_SETTING_IP_RESTRICT_UUID,
+        )
+        app_setting_ip_allowlist_obj = AppSetting.objects.get(
+            sodar_uuid=PROJECTROLES_APP_SETTING_IP_ALLOWLIST_UUID,
+        )
+
+        expected_ip_restrict = {
+            'name': 'ip_restrict',
+            'type': 'BOOLEAN',
+            'value': '0',
+            'value_json': {},
+            'sodar_uuid': uuid.UUID(PROJECTROLES_APP_SETTING_IP_RESTRICT_UUID),
+            'project': project_obj.id,
+            'app_plugin': None,
+            'user': None,
+            'user_modifiable': True,
+        }
+        expected_ip_allowlist = {
+            'name': 'ip_allowlist',
+            'type': 'STRING',
+            'value': '',
+            'value_json': {},
+            'sodar_uuid': uuid.UUID(PROJECTROLES_APP_SETTING_IP_ALLOWLIST_UUID),
+            'project': project_obj.id,
+            'app_plugin': None,
+            'user': None,
+            'user_modifiable': True,
+        }
+        app_setting_ip_restrict_dict = model_to_dict(
+            app_setting_ip_restrict_obj
+        )
+        app_setting_ip_allowlist_dict = model_to_dict(
+            app_setting_ip_allowlist_obj
+        )
+        app_setting_ip_restrict_dict.pop('id')
+        app_setting_ip_allowlist_dict.pop('id')
+
+        self.assertEqual(app_setting_ip_allowlist_dict, expected_ip_allowlist)
+        self.assertEqual(app_setting_ip_restrict_dict, expected_ip_restrict)
+
+        # Assert remote_data changes
+        expected = original_data
+        expected['users'][SOURCE_USER_UUID]['status'] = 'created'
+        expected['projects'][SOURCE_CATEGORY_UUID]['status'] = 'created'
+        expected['projects'][SOURCE_CATEGORY_UUID]['roles'][
+            SOURCE_CATEGORY_ROLE_UUID
+        ]['status'] = 'created'
+        expected['projects'][SOURCE_PROJECT_UUID]['status'] = 'created'
+        expected['projects'][SOURCE_PROJECT_UUID]['roles'][
+            SOURCE_PROJECT_ROLE_UUID
+        ]['status'] = 'created'
+
+        self.assertEqual(remote_data, expected)
+
+    def test_create_app_setting_local(self):
+        remote_data = self.default_data
+
+        remote_data['app_settings'][PROJECTROLES_APP_SETTING_IP_RESTRICT_UUID][
+            'local'
+        ] = True
+        remote_data['app_settings'][PROJECTROLES_APP_SETTING_IP_ALLOWLIST_UUID][
+            'local'
+        ] = True
+
+        original_data = deepcopy(remote_data)
+
+        self.remote_api.sync_source_data(self.source_site, remote_data)
 
         # Assert remote_data changes
         expected = original_data
@@ -892,12 +1001,33 @@ class TestSyncSourceData(
             level=SODAR_CONSTANTS['REMOTE_LEVEL_NONE'],
         )
 
+        # Init IP restrict setting
+        self._make_setting(
+            app_name='projectroles',
+            name='ip_restrict',
+            setting_type='BOOLEAN',
+            value=False,
+            project=project_obj,
+            sodar_uuid=PROJECTROLES_APP_SETTING_IP_RESTRICT_UUID,
+        )
+
+        # Init IP allowlist setting
+        self._make_setting(
+            app_name='projectroles',
+            name='ip_allowlist',
+            setting_type='STRING',
+            value='',
+            project=project_obj,
+            sodar_uuid=PROJECTROLES_APP_SETTING_IP_ALLOWLIST_UUID,
+        )
+
         # Assert preconditions
         self.assertEqual(Project.objects.all().count(), 2)
         self.assertEqual(RoleAssignment.objects.all().count(), 2)
         self.assertEqual(User.objects.all().count(), 2)
         self.assertEqual(RemoteProject.objects.all().count(), 3)
         self.assertEqual(RemoteSite.objects.all().count(), 2)
+        self.assertEqual(AppSetting.objects.count(), 2)
 
         remote_data = self.default_data
 
@@ -929,6 +1059,15 @@ class TestSyncSourceData(
 
         original_data = deepcopy(remote_data)
 
+        # Change projectroles app settings
+
+        remote_data['app_settings'][PROJECTROLES_APP_SETTING_IP_RESTRICT_UUID][
+            'value'
+        ] = True
+        remote_data['app_settings'][PROJECTROLES_APP_SETTING_IP_ALLOWLIST_UUID][
+            'value'
+        ] = '192.168.1.1'
+
         # Do sync
         self.remote_api.sync_source_data(self.source_site, remote_data)
 
@@ -938,6 +1077,7 @@ class TestSyncSourceData(
         self.assertEqual(User.objects.all().count(), 3)
         self.assertEqual(RemoteProject.objects.all().count(), 3)
         self.assertEqual(RemoteSite.objects.all().count(), 2)
+        self.assertEqual(AppSetting.objects.count(), 2)
 
         new_user = User.objects.get(username=new_user_username)
 
@@ -1062,6 +1202,47 @@ class TestSyncSourceData(
         peer_project_dict.pop('date_access')
         self.assertEqual(peer_project_dict, expected)
 
+        app_setting_ip_restrict_obj = AppSetting.objects.get(
+            sodar_uuid=PROJECTROLES_APP_SETTING_IP_RESTRICT_UUID,
+        )
+        app_setting_ip_allowlist_obj = AppSetting.objects.get(
+            sodar_uuid=PROJECTROLES_APP_SETTING_IP_ALLOWLIST_UUID,
+        )
+
+        expected_ip_restrict = {
+            'name': 'ip_restrict',
+            'type': 'BOOLEAN',
+            'value': '1',
+            'value_json': {},
+            'sodar_uuid': uuid.UUID(PROJECTROLES_APP_SETTING_IP_RESTRICT_UUID),
+            'project': project_obj.id,
+            'app_plugin': None,
+            'user': None,
+            'user_modifiable': True,
+        }
+        expected_ip_allowlist = {
+            'name': 'ip_allowlist',
+            'type': 'STRING',
+            'value': '192.168.1.1',
+            'value_json': {},
+            'sodar_uuid': uuid.UUID(PROJECTROLES_APP_SETTING_IP_ALLOWLIST_UUID),
+            'project': project_obj.id,
+            'app_plugin': None,
+            'user': None,
+            'user_modifiable': True,
+        }
+        app_setting_ip_restrict_dict = model_to_dict(
+            app_setting_ip_restrict_obj
+        )
+        app_setting_ip_allowlist_dict = model_to_dict(
+            app_setting_ip_allowlist_obj
+        )
+        app_setting_ip_restrict_dict.pop('id')
+        app_setting_ip_allowlist_dict.pop('id')
+
+        self.assertEqual(app_setting_ip_allowlist_dict, expected_ip_allowlist)
+        self.assertEqual(app_setting_ip_restrict_dict, expected_ip_restrict)
+
         # Assert update_data changes
         expected = original_data
         expected['users'][SOURCE_USER_UUID]['status'] = 'updated'
@@ -1071,6 +1252,181 @@ class TestSyncSourceData(
         expected['projects'][SOURCE_PROJECT_UUID]['roles'][new_role_uuid][
             'status'
         ] = 'created'
+        expected['app_settings'][PROJECTROLES_APP_SETTING_IP_RESTRICT_UUID][
+            'value'
+        ] = True
+        expected['app_settings'][PROJECTROLES_APP_SETTING_IP_ALLOWLIST_UUID][
+            'value'
+        ] = '192.168.1.1'
+
+        self.assertEqual(remote_data, expected)
+
+    def test_update_app_setting_local(self):
+        # Set up target category and project
+        category_obj = self._make_project(
+            title='NewCategoryTitle',
+            type=PROJECT_TYPE_CATEGORY,
+            parent=None,
+            description='New description',
+            readme='New readme',
+            sodar_uuid=SOURCE_CATEGORY_UUID,
+        )
+        project_obj = self._make_project(
+            title='NewProjectTitle',
+            type=PROJECT_TYPE_PROJECT,
+            parent=category_obj,
+            description='New description',
+            readme='New readme',
+            sodar_uuid=SOURCE_PROJECT_UUID,
+        )
+
+        # Set up user and roles
+        target_user = self._make_sodar_user(
+            username=SOURCE_USER_USERNAME,
+            name='NewFirstName NewLastName',
+            first_name='NewFirstName',
+            last_name='NewLastName',
+            email='newemail@example.com',
+        )
+        self._make_assignment(category_obj, target_user, self.role_owner)
+        self._make_assignment(project_obj, target_user, self.role_owner)
+
+        # Set up RemoteProject objects
+        self._make_remote_project(
+            project_uuid=category_obj.sodar_uuid,
+            project=category_obj,
+            site=self.source_site,
+            level=REMOTE_LEVEL_READ_ROLES,
+        )
+        self._make_remote_project(
+            project_uuid=project_obj.sodar_uuid,
+            project=project_obj,
+            site=self.source_site,
+            level=REMOTE_LEVEL_READ_ROLES,
+        )
+
+        # Set up Peer Objects
+        peer_site = RemoteSite.objects.create(
+            **{
+                'name': PEER_SITE_NAME,
+                'url': PEER_SITE_URL,
+                'mode': SITE_MODE_PEER,
+                'description': PEER_SITE_DESC,
+                'secret': None,
+                'sodar_uuid': PEER_SITE_UUID,
+                'user_display': PEER_SITE_USER_DISPLAY,
+            }
+        )
+
+        self._make_remote_project(
+            project_uuid=project_obj.sodar_uuid,
+            project=project_obj,
+            site=peer_site,
+            level=SODAR_CONSTANTS['REMOTE_LEVEL_NONE'],
+        )
+
+        # Init IP restrict setting
+        self._make_setting(
+            app_name='projectroles',
+            name='ip_restrict',
+            setting_type='BOOLEAN',
+            value=False,
+            project=project_obj,
+            sodar_uuid=PROJECTROLES_APP_SETTING_IP_RESTRICT_UUID,
+        )
+
+        # Init IP allowlist setting
+        self._make_setting(
+            app_name='projectroles',
+            name='ip_allowlist',
+            setting_type='STRING',
+            value='',
+            project=project_obj,
+            sodar_uuid=PROJECTROLES_APP_SETTING_IP_ALLOWLIST_UUID,
+        )
+
+        # Assert preconditions
+        self.assertEqual(Project.objects.all().count(), 2)
+        self.assertEqual(RoleAssignment.objects.all().count(), 2)
+        self.assertEqual(User.objects.all().count(), 2)
+        self.assertEqual(RemoteProject.objects.all().count(), 3)
+        self.assertEqual(RemoteSite.objects.all().count(), 2)
+        self.assertEqual(AppSetting.objects.count(), 2)
+
+        remote_data = self.default_data
+
+        # Change projectroles app settings
+
+        remote_data['app_settings'][PROJECTROLES_APP_SETTING_IP_RESTRICT_UUID][
+            'local'
+        ] = True
+        remote_data['app_settings'][PROJECTROLES_APP_SETTING_IP_ALLOWLIST_UUID][
+            'local'
+        ] = True
+        remote_data['app_settings'][PROJECTROLES_APP_SETTING_IP_RESTRICT_UUID][
+            'value'
+        ] = True
+        remote_data['app_settings'][PROJECTROLES_APP_SETTING_IP_ALLOWLIST_UUID][
+            'value'
+        ] = '192.168.1.1'
+
+        original_data = deepcopy(remote_data)
+
+        # Do sync
+        self.remote_api.sync_source_data(self.source_site, remote_data)
+
+        app_setting_ip_restrict_obj = AppSetting.objects.get(
+            sodar_uuid=PROJECTROLES_APP_SETTING_IP_RESTRICT_UUID,
+        )
+        app_setting_ip_allowlist_obj = AppSetting.objects.get(
+            sodar_uuid=PROJECTROLES_APP_SETTING_IP_ALLOWLIST_UUID,
+        )
+
+        expected_ip_restrict = {
+            'name': 'ip_restrict',
+            'type': 'BOOLEAN',
+            'value': '0',
+            'value_json': {},
+            'sodar_uuid': uuid.UUID(PROJECTROLES_APP_SETTING_IP_RESTRICT_UUID),
+            'project': project_obj.id,
+            'app_plugin': None,
+            'user': None,
+            'user_modifiable': True,
+        }
+        expected_ip_allowlist = {
+            'name': 'ip_allowlist',
+            'type': 'STRING',
+            'value': '',
+            'value_json': {},
+            'sodar_uuid': uuid.UUID(PROJECTROLES_APP_SETTING_IP_ALLOWLIST_UUID),
+            'project': project_obj.id,
+            'app_plugin': None,
+            'user': None,
+            'user_modifiable': True,
+        }
+        app_setting_ip_restrict_dict = model_to_dict(
+            app_setting_ip_restrict_obj
+        )
+        app_setting_ip_allowlist_dict = model_to_dict(
+            app_setting_ip_allowlist_obj
+        )
+        app_setting_ip_restrict_dict.pop('id')
+        app_setting_ip_allowlist_dict.pop('id')
+
+        self.assertEqual(app_setting_ip_allowlist_dict, expected_ip_allowlist)
+        self.assertEqual(app_setting_ip_restrict_dict, expected_ip_restrict)
+
+        # Assert update_data changes
+        expected = original_data
+        expected['users'][SOURCE_USER_UUID]['status'] = 'updated'
+        expected['projects'][SOURCE_CATEGORY_UUID]['status'] = 'updated'
+        expected['projects'][SOURCE_PROJECT_UUID]['status'] = 'updated'
+        expected['app_settings'][PROJECTROLES_APP_SETTING_IP_RESTRICT_UUID][
+            'value'
+        ] = '0'
+        expected['app_settings'][PROJECTROLES_APP_SETTING_IP_ALLOWLIST_UUID][
+            'value'
+        ] = ''
 
         self.assertEqual(remote_data, expected)